landing pages and drawer logic change
