+++ src/future-hopr-lib-components/Layout/drawer.tsx
@@ -13,6 +13,7 @@ import {
 } from '@mui/material';
 import { Link, useLocation } from 'react-router-dom';
 import { ApplicationMapType } from '../../router';
+import Details from '../../components/InfoBar/details';
 
 const drawerWidth = 240;
 const minDrawerWidth = 56;
@@ -97,6 +98,12 @@ const StyledListItemButton = styled(ListItemButton)`
   }
 ` as typeof ListItemButton;
 
+const SListItemIcon = styled(ListItemIcon)`
+  &.GroupIcon {
+    color: rgba(0, 0, 0, 0.2);
+  }
+`;
+
 type DrawerProps = {
   drawerItems: ApplicationMapType;
   drawerLoginState?: {
@@ -163,7 +170,7 @@ const Drawer = ({
                   placement="right"
                 >
                   <StyledListSubheader className="StyledListSubheader">
-                    <ListItemIcon className="ListItemIcon">{group.icon}</ListItemIcon>
+                    <SListItemIcon className="SListItemIcon GroupIcon">{group.icon}</SListItemIcon>
                   </StyledListSubheader>
                 </Tooltip>
               )
@@ -183,7 +190,7 @@ const Drawer = ({
                   onClick={handleButtonClick}
                   className="StyledListItemButton"
                 >
-                  <ListItemIcon className="ListItemIcon">{item.icon}</ListItemIcon>
+                  <SListItemIcon className="SListItemIcon">{item.icon}</SListItemIcon>
                   <ListItemText className="ListItemText">{item.name}</ListItemText>
                 </StyledListItemButton>
               </Tooltip>
@@ -191,6 +198,7 @@ const Drawer = ({
           </List>
         </div>
       ))}
+      {drawerVariant === 'temporary' && <Details style={{ margin: '0 auto 16px' }} />}
     </StyledDrawer>
   );
 };
