+++ src/future-hopr-lib-components/Layout/index.tsx
@@ -1,6 +1,6 @@
 /// <reference types="vite-plugin-svgr/client" />
 // Packages
-import React, { ReactNode, useState } from 'react';
+import React, { ReactNode, useState, useEffect } from 'react';
 import styled from '@emotion/styled';
 import { css } from '@emotion/react';
 import { Outlet } from 'react-router-dom';
@@ -81,14 +81,21 @@ const Layout: React.FC<{
   drawerRight,
   drawerType,
 }) => {
-  // Determine if the device is a mobile device based on the screen width
   const isMobile = useMediaQuery('(max-width: 500px)');
   const account = useAppSelector((store) => store.web3.account);
   const isConnected = useAppSelector((store) => store.auth.status.connected);
 
-  // Set the initial state of the drawer based on the device type
-  // If it's a mobile device, set the drawer to be closed by default
-  const [openedNavigationDrawer, set_openedNavigationDrawer] = useState(false);
+  const [openedNavigationDrawerPC, set_openedNavigationDrawerPC] = useState(false);
+  const [openedNavigationDrawerMobile, set_openedNavigationDrawerMobile] = useState(false);
+
+  const handleOpenedNavigationDrawer = (bool: boolean) => {
+    if (isMobile) set_openedNavigationDrawerMobile(bool);
+    else set_openedNavigationDrawerPC(bool);
+  };
+
+  useEffect(() => {
+    if (isConnected) set_openedNavigationDrawerPC(true);
+  }, [isConnected]);
 
   return (
     <SLayout className={`Layout${webapp ? ' webapp' : ''} ${className}`}>
@@ -98,21 +105,21 @@ const Layout: React.FC<{
         itemsNavbarRight={itemsNavbarRight}
         tallerNavBarOnMobile={tallerNavBarOnMobile}
         webapp={webapp}
-        set_openedNavigationDrawer={set_openedNavigationDrawer}
-        openedNavigationDrawer={openedNavigationDrawer}
+        set_openedNavigationDrawer={handleOpenedNavigationDrawer}
+        openedNavigationDrawer={isMobile ? openedNavigationDrawerMobile : openedNavigationDrawerPC}
       />
       {drawer && (
         <Drawer
           drawerType={drawerType}
           drawerItems={drawerItems}
           drawerLoginState={drawerLoginState}
-          set_openedNavigationDrawer={set_openedNavigationDrawer}
-          openedNavigationDrawer={openedNavigationDrawer}
+          set_openedNavigationDrawer={handleOpenedNavigationDrawer}
+          openedNavigationDrawer={isMobile ? openedNavigationDrawerMobile : openedNavigationDrawerPC}
         />
       )}
       <Content
         className={`Content ${drawerRight ? 'drawerRight' : ''}`}
-        openedNavigationDrawer={openedNavigationDrawer}
+        openedNavigationDrawer={isMobile ? openedNavigationDrawerMobile : openedNavigationDrawerPC}
         drawerRight={!!drawerRight}
       >
         {' '}
