--- src/sections/nodeLanding.tsx
@@ -1,16 +1,0 @@
-import Section from '../future-hopr-lib-components/Section';
-
-function NodeLandingPage() {
-  return (
-    <Section
-      className="Section--NodeLandingPage"
-      id="Section--NodeLandingPage"
-      yellow
-      fullHeightMin
-    >
-      Node Landing Page
-    </Section>
-  );
-}
-
-export default NodeLandingPage;
