+++ src/sections/staking-hub/noNodeAdded.tsx
@@ -1,9 +1,9 @@
 import { ReactNode } from 'react';
-import { useAppSelector } from '../store';
+import { useAppSelector } from '../../store';
 import styled from '@emotion/styled';
 
-import Button from '../future-hopr-lib-components/Button';
-import Section from '../future-hopr-lib-components/Section';
+import Button from '../../future-hopr-lib-components/Button';
+import Section from '../../future-hopr-lib-components/Section';
 import { Card, Chip, IconButton } from '@mui/material';
 import { Link } from 'react-router-dom';
