+++ src/sections/staking-hub/web3.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
-import { Store } from '../types/index';
+import { Store } from '../../types/index';
 
 //Stores
-import { useAppDispatch, useAppSelector } from '../store';
-import { web3Actions } from '../store/slices/web3';
-import { actionsAsync as web3ActionsAsync } from '../store/slices/web3/actionsAsync';
+import { useAppDispatch, useAppSelector } from '../../store';
+import { web3Actions } from '../../store/slices/web3';
+import { actionsAsync as web3ActionsAsync } from '../../store/slices/web3/actionsAsync';
 
 // HOPR Components
-import Section from '../future-hopr-lib-components/Section';
+import Section from '../../future-hopr-lib-components/Section';
 
-import ConnectWeb3 from '../components/ConnectWeb3';
+import ConnectWeb3 from '../../components/ConnectWeb3';
 
 function Section_Web3() {
   const dispatch = useAppDispatch();
